Move gateway setting to connectivity panel
